cbs/requirements.txt
aws-cdk-lib==2.140
aws-lambda-powertools==2.37
boto3~=1.34
constructs>=10.0.0,<11.0.0
pydantic[email]~=2.7
PyYAML~=6.0
